// SDKs/twilio.com/twilio_sync_v1/1.36.0/typescript/src/sdk/models/shared/syncv1servicesynclistsynclistpermission.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


export class SyncV1ServiceSyncListSyncListPermission extends SpeakeasyBase {
  @Metadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @Metadata({ data: "json, name=identity" })
  identity?: string;

  @Metadata({ data: "json, name=list_sid" })
  listSid?: string;

  @Metadata({ data: "json, name=manage" })
  manage?: boolean;

  @Metadata({ data: "json, name=read" })
  read?: boolean;

  @Metadata({ data: "json, name=service_sid" })
  serviceSid?: string;

  @Metadata({ data: "json, name=url" })
  url?: string;

  @Metadata({ data: "json, name=write" })
  write?: boolean;
}